Added FiniteDifferences.c Script and .gitignore file.
This script simulates the motion of a rope using the finite differences method. User can choose the parameters X and T at runtime using -X and -T in the command line. A display function was added in order to plot the position of all the points of the rope and see their time evolution.
.gitignore file contains a list of extensions which we do not want to include in the Git repository.